pyproject.toml
[project]
name = "rags-implementation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=0.5.0",
    "datasets>=2.18.0",
    "faiss-cpu>=1.8.0",
    "ipykernel>=6.0.0",
    "langchain>=0.3.0",
    "langchain-chroma>=0.1.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "llama-index>=0.11.0",
    "llama-index-core>=0.11.0",
    "llama-index-embeddings-huggingface>=0.3.0",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-llms-openai>=0.3.0",
    "llama-index-vector-stores-chroma>=0.2.0",
    "loguru>=0.7.0",
    "networkx>=3.0",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "ragas>=0.1.0",
    "rich>=13.0.0",
    "scikit-learn>=1.4.0",
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",
    "tqdm>=4.66.0",
    "transformers>=4.40.0",
    "umap-learn>=0.5.0",
]
